# Programming Tasks 05 to 08 - to be graded/Task 7/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Meeting Scheduler API
  description: >
    A simple API to schedule meetings, vote on slots, and analyze results.
    Implements a Design-First approach with strict status workflows (Draft -> Published).
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server

paths:
  
  # MEETING MANAGEMENT (Plan, propose, publish)
  
  /meetings:
    post:
      summary: Create a new meeting poll (Draft Mode)
      operationId: createMeeting
      description: >
        Creates a meeting in DRAFT status. Participants cannot vote until it is published.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingCreateRequest'
      responses:
        '201':
          description: Meeting created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /meetings/{meetingId}:
    get:
      summary: Get meeting details
      operationId: getMeeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Meeting details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
        '404':
          description: Meeting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /meetings/{meetingId}/publish:
    post:
      summary: Publish the meeting poll
      operationId: publishMeeting
      description: >
        Transitions the meeting status from DRAFT to PUBLISHED. 
        Requires at least one time slot to exist.
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Meeting published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
        '400':
          description: Meeting has no slots or is already closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Meeting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # SLOT MANAGEMENT (Add, remove slots)

  /meetings/{meetingId}/slots:
    post:
      summary: Add a new time slot to a meeting
      operationId: addSlot
      description: Can only be performed when meeting is in DRAFT mode.
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeSlotRequest'
      responses:
        '201':
          description: Slot added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlot'
        '400':
          description: Invalid time range or meeting is closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Meeting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /meetings/{meetingId}/slots/{slotId}:
    delete:
      summary: Remove a specific time slot
      operationId: removeSlot
      description: Can only be performed when meeting is in DRAFT mode.
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: slotId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Slot removed successfully
        '400':
          description: Cannot remove slot from published meeting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Slot or Meeting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # PARTICIPATION (Submit availability)

  /meetings/{meetingId}/submissions:
    post:
      summary: Submit availability for a meeting
      operationId: submitAvailability
      description: Only allowed for PUBLISHED meetings.
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionRequest'
      responses:
        '201':
          description: Submission accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
        '400':
          description: Invalid selection or meeting not open
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Meeting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    # ANALYSIS (Get submissions)

    get:
      summary: Get all submissions for a meeting (Analysis)
      operationId: getSubmissions
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of all participant submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Submission'
        '404':
          description: Meeting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    
    # DOMAIN OBJECTS

    Meeting:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: "DSE Project Kickoff"
        description:
          type: string
        status:
          type: string
          description: Current state of the poll workflow
          enum: [DRAFT, PUBLISHED, CLOSED]
          example: "DRAFT"
        slots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'

    TimeSlot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time

    Submission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        participantName:
          type: string
          example: "John Doe"
        selections:
          type: array
          items:
            $ref: '#/components/schemas/SlotSelection'

    SlotSelection:
      type: object
      properties:
        slotId:
          type: string
          format: uuid
        preference:
          type: string
          enum: [YES, NO, MAYBE]

    # REQUEST OBJECTS

    MeetingCreateRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
        initialSlots:
          type: array
          description: Optional initial slots. Must provide at least one before publishing.
          items:
            $ref: '#/components/schemas/TimeSlotRequest'

    TimeSlotRequest:
      type: object
      required: [startTime, endTime]
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time

    SubmissionRequest:
      type: object
      required: [participantName, selections]
      properties:
        participantName:
          type: string
          minLength: 1
        selections:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SlotSelection'
    
    # ERROR OBJECT

    Error:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: "Invalid input provided"